move execution_profile.rs to new execution module
