+++ docs/source/execution-profiles/maximal-example.md
@@ -9,8 +9,8 @@
 use scylla::query::Query;
 use scylla::speculative_execution::SimpleSpeculativeExecutionPolicy;
 use scylla::statement::{Consistency, SerialConsistency};
-use scylla::transport::ExecutionProfile;
-use scylla::transport::load_balancing::DefaultPolicy;
+use scylla::execution::ExecutionProfile;
+use scylla::execution::load_balancing::DefaultPolicy;
 use scylla::transport::retry_policy::FallthroughRetryPolicy;
 use std::{sync::Arc, time::Duration};
