+++ examples/execution_profile.rs
@@ -1,11 +1,11 @@
 use anyhow::Result;
+use scylla::execution::ExecutionProfile;
 use scylla::load_balancing;
 use scylla::query::Query;
 use scylla::retry_policy::{DefaultRetryPolicy, FallthroughRetryPolicy};
 use scylla::session::Session;
 use scylla::speculative_execution::PercentileSpeculativeExecutionPolicy;
 use scylla::statement::{Consistency, SerialConsistency};
-use scylla::transport::ExecutionProfile;
 use scylla::{SessionBuilder, SessionConfig};
 use std::env;
 use std::sync::Arc;
