+++ scylla/src/execution/mod.rs
@@ -1,3 +1,8 @@
 //! This module holds entities that control, customize, trace, and measure statement execution.
 //! This includes:
+//! - execution profiles,
 //! - TODO.
+
+pub mod execution_profile;
+
+pub use execution_profile::ExecutionProfile;
