+++ scylla/src/lib.rs
@@ -277,12 +277,12 @@ pub use statement::query;
 #[allow(deprecated)]
 pub use frame::response::cql_to_rust::{self, FromRow};
 
+pub use execution::execution_profile::ExecutionProfile;
 #[allow(deprecated)]
 pub use session::{
     CachingSession, GenericCachingSession, IntoTypedRows, LegacyCachingSession, LegacySession,
     Session, SessionBuilder, SessionConfig,
 };
-pub use transport::execution_profile::ExecutionProfile;
 #[allow(deprecated)]
 pub use transport::legacy_query_result::LegacyQueryResult;
 pub use transport::query_result::{QueryResult, QueryRowsResult};
@@ -291,7 +291,7 @@ pub use transport::query_result::{QueryResult, QueryRowsResult};
 pub use session::CloudSessionBuilder;
 
 pub use cluster::host_filter;
-pub use transport::execution_profile;
+pub use execution::execution_profile;
 pub use transport::load_balancing;
 pub use transport::retry_policy;
 pub use transport::speculative_execution;
