+++ scylla/src/session/session_builder.rs
@@ -1,11 +1,11 @@
 //! SessionBuilder provides an easy way to create new Sessions
 
 use crate::connection::{AddressTranslator, Compression, SelfIdentity};
+use crate::execution::execution_profile::ExecutionProfileHandle;
 #[allow(deprecated)]
 use crate::session::{
     CurrentDeserializationApi, GenericSession, LegacyDeserializationApi, SessionConfig,
 };
-use crate::transport::execution_profile::ExecutionProfileHandle;
 
 #[cfg(feature = "cloud")]
 use crate::cloud::{CloudConfig, CloudConfigError};
@@ -1035,10 +1035,11 @@ mod tests {
     use scylla_cql::frame::types::SerialConsistency;
     use scylla_cql::Consistency;
 
-    use super::{Compression, SessionBuilder};
+    use super::SessionBuilder;
     use crate::cluster::node::KnownNode;
+    use crate::connection::Compression;
+    use crate::execution::execution_profile::{defaults, ExecutionProfile};
     use crate::test_utils::setup_tracing;
-    use crate::transport::execution_profile::{defaults, ExecutionProfile};
     use std::net::{IpAddr, Ipv4Addr, SocketAddr};
     use std::time::Duration;
