+++ scylla/src/transport/mod.rs
@@ -1,6 +1,5 @@
 pub mod downgrading_consistency_retry_policy;
 pub mod errors;
-pub mod execution_profile;
 pub mod iterator;
 pub mod legacy_query_result;
 pub mod load_balancing;
@@ -12,5 +11,4 @@ pub mod retry_policy;
 pub mod speculative_execution;
 
 pub use crate::frame::Authenticator;
-pub use execution_profile::ExecutionProfile;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
